add boot_client parameter to bos_inst
Currently, *all* attempts to perform a bos_inst operation will cause
the NIM server to attempt to reboot the client. This results in a long
delay when the client doesn't exist yet, and prevents staging a bos_inst
operation for a later time.

The module relies on the default behavior of the NIM server (which I
think is correct), but needs a way to pass the attribute to alter that
behavior.

The change uses a Truthy default value for boot_client to preserve the
current behavior when it is not specified.

Signed-off-by: Stephen Ulmer <ulmer@ulmer.org>