Display in flrtvc module a message if no interim fixes were installed. "There are no vulnerabilities."
Problem: The flrtvc module will fail if a list with no vulnerabilities is processed.

This also will display a message indicating that there are no interim fixes
in the package downloaded.
Example, if the flrtvc downloads filesets for Java updates but
this fixes do not come in the epkg format. The user can still uncompress manually
the package and choose which version of fixes to be installed.

The message for this situaiton will be "There are no interim fixes in epkg formate to be installed"