Localpatchserver patch: (#523)
* Localpatchserver patch:
Includes the abillity to local clone in an http server the packages
from ibm ifix  thus avoiding the usage of internet/proxy in the aix
systems. The method is simple you specify your server and the path of
the patches and the flrtvc module replaces the urls from flrtvc.ksh to
the desired local ones. It will need of course to be set with apar_csv
in order to find a local copy of apar.csv. See demo playbooks for this.

Signed-off-by: ptsiamis@gmail.com <ptsiamis@gmail.com>

* Updated demo localpatch playbooks to include fqcn
Updated flrtvc.py to comply with pep8 standarts

Signed-off-by: ptsiamis@gmail.com <ptsiamis@gmail.com>

* Fixed flake8 issue: plugins/modules/flrtvc.py:1063:32: W504 line break after binary operator

Signed-off-by: ptsiamis@gmail.com <ptsiamis@gmail.com>

---------

Signed-off-by: ptsiamis@gmail.com <ptsiamis@gmail.com>